unsigned long to int replacement on various types.
2000-12-12  Eskil Heyn Olsen  <eskil@eazel.com>

	* components/services/install/lib/eazel-package-system-private.h:
	* components/services/install/lib/eazel-package-system-rpm3-private
	.h:
	* components/services/install/lib/eazel-package-system-rpm3.c:
	(rpm_packagedata_fill_from_rpm_header),
	(rpm_packagedata_fill_from_file), (rpm_packagedata_new_from_file),
	(eazel_package_system_rpm3_load_package),
	(eazel_package_system_rpm3_query_impl),
	(eazel_package_system_rpm3_query_substr),
	(eazel_package_system_rpm3_query_requires),
	(eazel_package_system_rpm3_query):
	* components/services/install/lib/eazel-package-system-skeleton.c:
	(eazel_package_system_skeleton_query):
	* components/services/install/lib/eazel-package-system-types.c:
	(packagedata_new):
	* components/services/install/lib/eazel-package-system-types.h:
	* components/services/install/lib/eazel-package-system.c:
	(eazel_package_system_load_package), (eazel_package_system_query):
	* components/services/install/lib/eazel-package-system.h:
	unsigned long to int replacement on various types.

+++ components/services/install/lib/eazel-package-system-private.h
@@ -47,12 +47,12 @@ typedef EazelPackageSystem*(*EazelPackageSystemConstructorFunc) (GList*);
 typedef	PackageData* (*EazelPackageSytemLoadPackageFunc) (EazelPackageSystem*,
 							  PackageData*, 
 							  const char*, 
-							  unsigned long);
+							  int);
 typedef GList* (*EazelPackageSytemQueryFunc) (EazelPackageSystem*, 
 					      const char*, 
 					      gpointer key, 
 					      EazelPackageSystemQueryEnum, 
-					   unsigned long);
+					      int);
 typedef void (*EazelPackageSytemInstallFunc) (EazelPackageSystem*, 
 					      const char*,
 					      GList*,
